carga/requirements.txt
psycopg2==2.9.3
pandas==1.3.3
sqlalchemy==1.4.25
python-dotenv==0.19.2